Remove Var.typeOf
Motivation: Var is currently not a very useful typeclass. The fact
that it allows creating variables from a "type" and also getting back
the "type" from a variable limits what a variable can be-it has to
have one of the predefined types. Removing the ability to query the type
of a variable makes more Var instances possible.